Learned classifier: identity-initialised residual support head
The parameter-free vote is already strong, so the learned head is confined to the two
regimes it provably cannot enter. Diagnostics established both: at k=1 every readout is
mathematically identical and on wrong predictions the correct support is in the top-5
92.5% of the time, so only a second evidence source can help; at k>=8, 15.6% of wrong
predictions had the correct label's best support ranked first and were outvoted, which
is a voting failure a per-support reweighting targets directly.

One scoring rule for every k. Centre the episode on its support mean, take
differentiable neighbours as the base score, add a learned scalar residual per support
and per candidate, and a frozen-SBERT label-text term weighted by how many supports a
candidate has. The residual heads are zero-initialised and lambda(k>0) starts at zero,
so the untrained classifier equals the control bit-for-bit; a candidate with no supports
is scored by text alone, which makes k=0 the same code path rather than a second head.

Training adds per-candidate support masking so some candidates must be recognised from
their name, which is the mechanism-level guard against memorising the ~90 training
labels - the failure that sank every previous text head here.

Centring is the classifier's own first step, not a shared readout: neighbors.py and the
1-NN/prototype/ridge controls are untouched and stay identical for every provider.

+++ model/support/token_mixer.py
@@ -206,7 +206,7 @@ class SupportTokenMixer(nn.Module):
                 candidate_mask=candidate_mask.index_select(0, zero_rows),
                 candidate_slot=candidate_slot.index_select(0, zero_rows),
             )
-            logits.index_copy_(0, zero_rows, result["logits"])
+            logits.index_copy_(0, zero_rows, result["logits"].to(logits.dtype))
         if len(few_rows):
             if any(value is None for value in (support_feature, support_label_text, support_bound,
                                                support_mask, support_pair_slot)):
@@ -222,9 +222,9 @@ class SupportTokenMixer(nn.Module):
                 support_pair_slot=support_pair_slot.index_select(0, few_rows),
                 candidate_mask=candidate_mask.index_select(0, few_rows),
             )
-            logits.index_copy_(0, few_rows, result["logits"])
+            logits.index_copy_(0, few_rows, result["logits"].to(logits.dtype))
             support_weight = query_feature.new_zeros(support_mask.shape, dtype=torch.float32)
-            support_weight.index_copy_(0, few_rows, result["support_weight"])
+            support_weight.index_copy_(0, few_rows, result["support_weight"].to(support_weight.dtype))
         return {"logits": logits, "support_weight": support_weight}
 
     def telemetry(self) -> dict[str, float]:
